+++ pytorch_lightning/accelerators/cpu.py
@@ -33,6 +33,3 @@ class CPUAccelerator(Accelerator):
             raise MisconfigurationException(f"Device should be CPU, got {self.root_device} instead.")
 
         return super().setup(trainer)
-
-    def get_device_stats(self, device: Union[str, torch.device]) -> Dict[str, Any]:
-        raise NotImplementedError